+++ BUILD.gn
@@ -151,12 +151,10 @@ template("glslang_sources_common") {
       "SPIRV/SpvBuilder.h",
       "SPIRV/SpvPostProcess.cpp",
       "SPIRV/SpvTools.h",
-      "SPIRV/bitutils.h",
       "SPIRV/disassemble.cpp",
       "SPIRV/disassemble.h",
       "SPIRV/doc.cpp",
       "SPIRV/doc.h",
-      "SPIRV/hex_float.h",
       "SPIRV/spirv.hpp11",
       "SPIRV/spvIR.h",
       "SPIRV/spvUtil.h",
@@ -173,7 +171,9 @@ template("glslang_sources_common") {
       "glslang/Include/SpirvIntrinsics.h",
       "glslang/Include/Types.h",
       "glslang/Include/arrays.h",
+      "glslang/Include/bitutils.h",
       "glslang/Include/defer.h",
+      "glslang/Include/hex_float.h",
       "glslang/Include/intermediate.h",
       "glslang/Include/visibility.h",
       "glslang/MachineIndependent/Constant.cpp",
