+++ SPIRV/CMakeLists.txt
@@ -44,13 +44,11 @@ set(SPIRV_SOURCES
     PARENT_SCOPE)
 
 set(SPIRV_HEADERS
-    ${CMAKE_CURRENT_SOURCE_DIR}/bitutils.h
     ${CMAKE_CURRENT_SOURCE_DIR}/spirv.hpp11
     ${CMAKE_CURRENT_SOURCE_DIR}/GLSL.std.450.h
     ${CMAKE_CURRENT_SOURCE_DIR}/GLSL.ext.EXT.h
     ${CMAKE_CURRENT_SOURCE_DIR}/GLSL.ext.KHR.h
     ${CMAKE_CURRENT_SOURCE_DIR}/GlslangToSpv.h
-    ${CMAKE_CURRENT_SOURCE_DIR}/hex_float.h
     ${CMAKE_CURRENT_SOURCE_DIR}/Logger.h
     ${CMAKE_CURRENT_SOURCE_DIR}/SpvBuilder.h
     ${CMAKE_CURRENT_SOURCE_DIR}/spvIR.h
