+++ SPIRV/SpvBuilder.cpp
@@ -48,7 +48,7 @@
 
 #include "SpvBuilder.h"
 #include "spvUtil.h"
-#include "hex_float.h"
+#include "glslang/Include/hex_float.h"
 
 #ifndef _WIN32
     #include <cstdio>
